% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentoapp.R
\name{sento_app}
\alias{sento_app}
\title{Shiny application for sentiment computation, aggregation and visualisation}
\usage{
sento_app()
}
\description{
A Shiny application built on the \pkg{sentometrics} package to allow straightforward
textual sentiment computation, aggregation and visualisation. The default loaded .csv file is a
subset of the \code{\link[sentometrics]{usnews}} corpus.
}
\details{
This Shiny application demonstrates mainly the \code{\link[sentometrics]{compute_sentiment}}, 
\code{\link[sentometrics]{corpus_summarize}} and \code{\link[sentometrics]{aggregate.sentiment}}
functions from the \pkg{sentometrics} package, allowing to gain insights into an uploaded corpus. The corpus should 
be uploaded in .csv format. Lexicons and valence shifters can be chosen from the built-in options or 
uploaded, and the weighting schemes are those available within \pkg{sentometrics}. All calculated values 
and statistics can be downloaded as a .csv file.
}
\author{
Samuel Borms, Jeroen Van Pelt
}
