% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentometrics.app.R
\docType{package}
\name{sentometrics.app-package}
\alias{sentometrics.app}
\alias{sentometrics.app-package}
\title{sentometrics.app: A Shiny Visual Interface to the sentometrics R Package}
\description{
The \pkg{sentometrics.app} package offers a Shiny interface for fast document-level textual
sentiment computation and aggregation into textual sentiment time series. The obtained values can be downloaded 
for further use. The functionalities provided are partly those from the \pkg{sentometrics} package.

Use the \code{\link{sento_app}} function to run the application.
}
\note{
Please cite the package in publications. Use \code{citation("sentometrics.app")}.
}
\references{
Ardia, Bluteau, Borms and Boudt (2019). ``The R package sentometrics to compute, aggregate and
predict with textual sentiment''. \emph{Working paper}, \url{https://doi.org/10.2139/ssrn.3067734}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sborms/sentometrics.app}
  \item Report bugs at \url{https://github.com/sborms/sentometrics.app/issues}
}

}
\author{
\strong{Maintainer}: Samuel Borms \email{samuel.borms@unine.ch}

Authors:
\itemize{
  \item Jeroen Van Pelt \email{jeroenvanpelt@hotmail.com}
}

Other contributors:
\itemize{
  \item Andres Algaba \email{andres.algaba@vub.be} [contributor]
}

}
